initial work on PoC for message spooling using kafka's Log
reworked GELF message inputs to use the new journalbuffer, from where every message is persisted in a very simple, unparsed form to a kafka log file.
currently there is no reader part, that is the next step

this code is intended as a proof of concept for the code structure, functionality as well as setting a baseline performance.
an in-memory journal (i.e. a non-journalling journal) is forthcoming to measure the overhead of disk access, as well as providing a low latency alternative in cases where persistence and guarding against overwhelming message peaks are not needed.